Combine and rename modules so it makes more sense
