+++ .gitignore
@@ -129,3 +129,4 @@ fabric.properties
 # Project specific
 build/result.json
 build/notes.txt
+src/__init__.py
