--- scrapper/spider/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["domica", "eentweedriewonen", "eervast", "nederwoon"]
